Create core.magics.pylab according to new API.
